server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt[crypto]>=4.3.0",
    "fastapi>=0.115.11",
    "pydantic-settings>=2.8.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "python-socketio>=5.13.0",
    "sqladmin[full]>=0.20.1",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]
